+++ cmd/test/flags.go
